MOVE: update import statements on moved files
