+++ API/src/main/java/com/litcode/server/dto/UserProfile.java
@@ -1,4 +1,4 @@
-package com.litcode.server.payload;
+package com.litcode.server.dto;
 
 import java.io.Serializable;
 import java.util.List;
